Diagonal movement implemented

+++ SlowpokeHubs/SlowpokeHub.cs
@@ -48,7 +48,7 @@ namespace SlowpokeHubs
             return MechanicEngine.ViewPort.GetActiveBodies(_connectionsPlayerMapping[Context.ConnectionId]).ToList();
 		}
 
-		public void MoveBody(int x, int y)
+		public void MoveBody(double x, double y)
 		{
             var player = MechanicEngine.GetPlayerBody(_connectionsPlayerMapping[Context.ConnectionId]);
